requirements.txt
fastapi 
uvicorn
transformers 
torch
pdfminer.six 
python-docx
nltk
flashtext
dotenv
fpdf